requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
backcall==0.2.0
black==23.1.0
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
configargparse==1.5.3
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
dm-tree==0.1.8
easydict==1.10
einops==0.6.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
flask==2.2.3
fonttools==4.39.0
frozenlist==1.3.3
ftfy==6.1.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
# gym==0.21.0
huggingface-hub==0.13.1
hydra-core==1.3.2
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.6.10
lazy-loader==0.2
markdown==3.4.1
markupsafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.16.0
open3d==0.17.0
openai==1.8.0
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
platformdirs==3.1.1
plotly==5.13.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pygments==2.14.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pywavelets==1.4.1
pyyaml==6.0
pyzmq==25.0.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
sentencepiece==0.1.97
six==1.16.0
stack-data==0.6.2
submitit==1.4.5
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.12
tokenizers==0.13.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
transforms3d==0.4.1
typing-extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
werkzeug==2.2.3
widgetsnbextension==4.0.5
yarl==1.8.2
zipp==3.15.0
sentence-transformers
peft
faiss-gpu
rank_bm25